move some function to util.py
