+++ util.py
@@ -1,12 +1,12 @@
 import numpy as np
 import torch
 import torch.nn.functional as F
-import math
-from collections import OrderedDict
 import os
-from scipy.ndimage import morphology
 from skimage.io import imsave
 import cv2
+from PIL import Image
+import matplotlib.pyplot as plt
+import torchvision.transforms as transforms
 
 
 def dict2obj(d):
@@ -311,3 +311,79 @@ def save_obj(filename, vertices, faces, textures=None, uvcoords=None, uvfaces=No
         with open(filename_mtl, 'w') as f:
             f.write('newmtl %s\n' % material_name)
             f.write('map_Kd %s\n' % os.path.basename(filename_texture))
+
+
+def face_seg(img, net, cropped_size):
+    face_area = [1, 2, 3, 4, 5, 6, 10, 11, 12, 13]
+    to_tensor = transforms.Compose([
+        transforms.ToTensor(),
+        transforms.Normalize((0.485, 0.456, 0.406), (0.229, 0.224, 0.225)),
+    ])
+    pil_image = Image.fromarray(cv2.cvtColor(img, cv2.COLOR_BGR2RGB))
+    resize_pil_image = pil_image.resize((512, 512), Image.BILINEAR)
+    tensor_image = to_tensor(resize_pil_image)
+    tensor_image = torch.unsqueeze(tensor_image, 0)
+    tensor_image = tensor_image.cuda()
+    out = net(tensor_image)[0]
+    parsing = out.squeeze(0).cpu().detach().numpy().argmax(0)
+    vis_parsing_anno = parsing.copy().astype(np.uint8)
+    vis_parsing_anno_color = np.zeros((vis_parsing_anno.shape[0], vis_parsing_anno.shape[1]))
+    num_of_class = np.max(vis_parsing_anno)
+
+    for pi in range(1, num_of_class + 1):
+        if pi in face_area:
+            index = np.where(vis_parsing_anno == pi)
+            vis_parsing_anno_color[index[0], index[1]] = 1
+    image_mask = cv2.resize(vis_parsing_anno_color, (cropped_size, cropped_size))
+    image_mask = image_mask[..., None].astype('float32')
+    image_mask = image_mask.transpose(2, 0, 1)
+    image_mask_bn = np.zeros_like(image_mask)
+    image_mask_bn[np.where(image_mask != 0)] = 1.
+
+    return image_mask_bn[None, :, :, :]
+
+
+def crop_img(ori_image, rect, cropped_size):
+    l, t, r, b = rect
+    center_x = r - (r - l) // 2
+    center_y = b - (b - t) // 2
+    w = (r - l) * 1.2
+    h = (b - t) * 1.2
+    crop_size = max(w, h)
+    if crop_size > cropped_size:
+        crop_ly = int(max(0, center_y - crop_size // 2))
+        crop_lx = int(max(0, center_x - crop_size // 2))
+        crop_ly = int(min(ori_image.shape[0] - crop_size, crop_ly))
+        crop_lx = int(min(ori_image.shape[1] - crop_size, crop_lx))
+        crop_image = ori_image[crop_ly: int(crop_ly + crop_size), crop_lx: int(crop_lx + crop_size), :]
+    else:
+
+        crop_ly = int(max(0, center_y - cropped_size // 2))
+        crop_lx = int(max(0, center_x - cropped_size // 2))
+        crop_ly = int(min(ori_image.shape[0] - cropped_size, crop_ly))
+        crop_lx = int(min(ori_image.shape[1] - cropped_size, crop_lx))
+        crop_image = ori_image[crop_ly: int(crop_ly + cropped_size), crop_lx: int(crop_lx + cropped_size), :]
+    new_rect = [l - crop_lx, t - crop_ly, r - crop_lx, b - crop_ly]
+    return crop_image, new_rect
+
+
+def resize_para(ori_frame):
+    w, h, c = ori_frame.shape
+    d = max(w, h)
+    scale_to = 640 if d >= 1280 else d / 2
+    scale_to = max(64, scale_to)
+    input_scale = d / scale_to
+    w = int(w / input_scale)
+    h = int(h / input_scale)
+    image_info = [w, h, input_scale]
+    return image_info
+
+
+def draw_train_process(title, iters, loss, label_loss):
+    plt.title(title, fontsize=24)
+    plt.xlabel("iter", fontsize=20)
+    plt.ylabel("loss", fontsize=20)
+    plt.plot(iters, loss, color='red', label=label_loss)
+    plt.legend()
+    plt.grid()
+    plt.show()
